Refactors files, OpticLang was a mess
